requirements.txt
openai==1.35.14
python-dotenv==1.0.1
tqdm==4.66.4
datasketch==1.6.5
numpy==1.26.4
pandas==2.2.2
requests==2.32.3
langchain==0.2.12
sqlglot==25.10.0